Some refactoring for easier navigation in files
